+++ react-client-app/src/components/MappingTbl.jsx
@@ -18,7 +18,7 @@ import {
 } from "@chakra-ui/react"
 
 import { ArrowForwardIcon } from '@chakra-ui/icons'
-import { useGet } from '../api/values'
+import { useGet, usePost } from '../api/values'
 import ConceptTag from './ConceptTag'
 import MappingModal from './MappingModal'
 import AnalysisModal from './AnalysisModal'
@@ -63,25 +63,28 @@ const MappingTbl = () => {
             })
     }, []);
 
-    useEffect(() => {
+    useEffect(async () => {
         // run when map diagram state has changed
         if (!mapDiagram.image) {
             // if no map diagram is loaded, request to get a new one
-            window.getSVG().then(diagram => {
-                setMapDiagram(mapDiagram => ({ ...mapDiagram, image: diagram.getElementsByTagName("svg")[0] }))
-                if (svg.current) {
-                    if (svg.current.hasChildNodes()) {
-                        // remove all other diagrams if they exist
-                        while (svg.current.firstChild) {
-                            svg.current.removeChild(svg.current.lastChild);
-                        }
+            
+            const result = await usePost(window.location.href, { 'get_svg': true }, false);
+            const diagramString = await result.data
+            var parser = new DOMParser();
+            var diagram = parser.parseFromString(diagramString, "text/html");
+
+            setMapDiagram((mapDiagram2) => ({ ...mapDiagram2, image: diagram.getElementsByTagName("svg")[0] }));
+            if (svg.current) {
+                if (svg.current.hasChildNodes()) {
+                    while (svg.current.firstChild) {
+                        svg.current.removeChild(svg.current.lastChild);
                     }
-                    svg.current.appendChild(diagram.getElementsByTagName("svg")[0])
                 }
-                if (downLoadingImgRef.current == true) {
-                    downloadImage(diagram.getElementsByTagName("svg")[0])
-                }
-            })
+                svg.current.appendChild(diagram.getElementsByTagName("svg")[0]);
+            }
+            if (downLoadingImgRef.current == true) {
+                downloadImage(diagram.getElementsByTagName("svg")[0]);
+            }
         }
         else {
             if (svg.current) {
@@ -98,46 +101,50 @@ const MappingTbl = () => {
 
 
     // call refresh rules function from django then get new data
-    const refreshRules = () => {
+    const refreshRules = async () => {
         setLoading(true)
         setLoadingMessage("Refreshing rules")
-        window.refreshRules().then(res => {
+        try {
+            await usePost(window.location.href, { 'refresh_rules': true }, false)
             setLoadingMessage("Rules Refreshed. Getting Mapping Rules")
             window.location.reload(true)
-        })
-            .catch(error => {
-                console.log(error)
-            })
+        }
+        catch (err) {
+            console.log(err)
+        }
+
     }
     // download map diagram
     const downloadImage = (img) => {
-        setDownloadingImg(true)
-        // if the image has been loaded then download it, otherwise, wait until image has been loaded
-        // then call the function again
+        setDownloadingImg(true);
         if (mapDiagram.image || img) {
-            let svg
-            if (img) { svg = img }
-            else { svg = mapDiagram.image }
-            // download the image then 
-            window.downloadImage(svg).then(res => {
-                setDownloadingImg(false)
-                downLoadingImgRef.current = false
-            })
+          let svg2;
+          if (img) {
+            svg2 = img;
+          } else {
+            svg2 = mapDiagram.image;
+          }
+          var serializer = new XMLSerializer();
+          var source = serializer.serializeToString(svg2);
+          if (!source.match(/^<svg[^>]+xmlns="http\:\/\/www\.w3\.org\/2000\/svg"/)) {
+            source = source.replace(/^<svg/, '<svg xmlns="http://www.w3.org/2000/svg"');
+          }
+          if (!source.match(/^<svg[^>]+"http\:\/\/www\.w3\.org\/1999\/xlink"/)) {
+            source = source.replace(/^<svg/, '<svg xmlns:xlink="http://www.w3.org/1999/xlink"');
+          }
+          source = '<?xml version="1.0" standalone="no"?>\r\n' + source;
+          var url = "data:image/svg+xml;charset=utf-8," + encodeURIComponent(source);
+          var element = document.createElement("a");
+          element.download = "diagram.svg";
+          element.href = url;
+          element.click();
+          element.remove();
+          setDownloadingImg(false);
+          downLoadingImgRef.current = false;
+        } else {
+          downLoadingImgRef.current = true;
         }
-        else {
-            // used to check if the image is waiting to be downloaded when the image is retrieved
-            downLoadingImgRef.current = true
-        }
-    }
-    // remove the map diagram from html
-    const removeDiagram = () => {
-        if (svg.current) {
-            if (svg.current.hasChildNodes()) {
-                svg.current.removeChild(mapDiagram.image)
-            }
-        }
-        setMapDiagram(mapDiagram => ({ ...mapDiagram, image: null }))
-    }
+      };
 
     // apply destination table and source table filters to data
     const applyFilters = (variable) => {
@@ -157,7 +164,7 @@ const MappingTbl = () => {
         if (filters.find(filter => filter.name == value) == null) {
             setFilters(current => [...current, { title: "Destination Table:", name: value }])
             setDestinationTableFilter(current => [...current, value])
-            //removeDiagram()
+            
         }
     };
     // if filter does not already exist, create a new source table filter
@@ -166,14 +173,14 @@ const MappingTbl = () => {
         if (filters.find(filter => filter.name == value) == null) {
             setFilters(current => [...current, { title: "Source Table:", name: value }])
             setSourceTableFilter(current => [...current, value])
-            //removeDiagram()
+            
         }
 
     };
     // remove a filter. Called inside concept tag
     const removeFilter = (title, name) => {
         setFilters(current => current.filter(filter => filter.name != name || filter.title != title))
-        //removeDiagram()
+        
         if (title.includes("Destination Table")) {
             setDestinationTableFilter(current => current.filter(filter => filter != name))
         }
@@ -212,8 +219,8 @@ const MappingTbl = () => {
         <div >
             <MappingModal isOpen={isOpen} onOpen={onOpen} onClose={onClose}>
                 <SummaryTbl values={values}
-                filters={filters}removeFilter={removeFilter} setDestinationFilter={setDestinationFilter}setSourceFilter={setSourceFilter}
-                destinationTableFilter={destinationTableFilter}sourceTableFilter={sourceTableFilter}/>
+                    filters={filters} removeFilter={removeFilter} setDestinationFilter={setDestinationFilter} setSourceFilter={setSourceFilter}
+                    destinationTableFilter={destinationTableFilter} sourceTableFilter={sourceTableFilter} />
             </MappingModal>
             <AnalysisModal isOpenAnalyse={isOpenAnalyse} onOpenAnalyse={onOpenAnalyse} onCloseAnalyse={onCloseAnalyse}>
                 <ConceptAnalysis data={data} values={values} filters={filters}/>
@@ -227,7 +234,7 @@ const MappingTbl = () => {
                 <Button variant="blue" onClick={onOpen}>Show Summary view</Button>
                 <Button variant="blue" onClick={onOpenAnalyse}>Analyse Rules</Button>
             </HStack>
-            
+
             <div>
                 {mapDiagram.showing &&
                     <>
@@ -253,9 +260,9 @@ const MappingTbl = () => {
                 <div>{error}</div>
                 :
                 <RulesTbl values={values}
-                filters={filters}removeFilter={removeFilter} setDestinationFilter={setDestinationFilter}setSourceFilter={setSourceFilter}
-                destinationTableFilter={destinationTableFilter}sourceTableFilter={sourceTableFilter} applyFilters={applyFilters}/>
-                
+                    filters={filters} removeFilter={removeFilter} setDestinationFilter={setDestinationFilter} setSourceFilter={setSourceFilter}
+                    destinationTableFilter={destinationTableFilter} sourceTableFilter={sourceTableFilter} applyFilters={applyFilters} />
+
             }
         </div>
     );